Question policy

+++ app/Http/Controllers/QuestionController.php
@@ -10,19 +10,11 @@ use Illuminate\Support\Facades\DB;
 class QuestionController extends Controller
 {
 
-    public function index()
-    {
-        $questions = Question::all();
-        return view('question.show', compact('questions'));
-    }
-
-    public function create(){
-        return redirect('/');
-    }
-
     public function store(QuestionRequest $request)
     {
         $validated = $request->validated();
+        $this->authorize('create');
+
         Question::create($validated);
         $id = $validated["test_id"];
         return redirect("/test/$id");
@@ -31,24 +23,28 @@ class QuestionController extends Controller
 
     public function show(Question $question)
     {
-        return view('question.edit');
-        return $question;
+        $this->authorize('view', $question);
+        return view('question.edit', compact('question'));
     }
 
     public function update(QuestionRequest $request, Question $question)
     {
         $validated = $request->validated();
+        $this->authorize('update', $question);
         $question->update($validated);
         return $question;
     }
 
     public function destroy(Question $question)
     {
+        $this->authorize('delete', $question);
         $question->delete();
     }
 
     public function optionCount($id) {
         $question = Question::findOrFail($id);
+        $this->authorize('view', $question);
+
         $test = Test::findOrFail($question->test_id);
         $results = [];
 
@@ -59,4 +55,15 @@ class QuestionController extends Controller
             ->get();
         return $results;
     }
+
+    /*public function index()
+    {
+
+        $questions = Question::all();
+        return view('question.show', compact('questions'));
+    }*/
+
+    /*public function create(){
+       return redirect('/');
+   }*/
 }
